requirements-base.txt
# Tools
ipython


# General libs
pytz
bleach
Markdown
Pillow
psycopg2


# Django
Django
django-authtools
django-braces
django-extensions
social-auth-core[openidconnect]
social-auth-app-django
requests-oauthlib
django-bootstrap3

